+++ sudoku/sudoku.hpp
@@ -4,16 +4,16 @@
 #include <iostream>
 
 class Sudoku {
-    std::vector<std::vector<size_t>> grid;
-    std::vector<std::vector<bool> > fixed;
+    std::vector<std::vector<int16_t>> grid;
+    std::vector<std::vector<bool>> fixed;
 
 public:
     Sudoku(size_t n);
 
-    void set_position(size_t i, size_t j, size_t value);
-    void set_row(size_t i, std::vector<size_t>&& value);
+    void set_position(size_t i, size_t j, int16_t value);
+    void set_row(size_t i, std::vector<int16_t>&& value);
 
-    size_t get_value(size_t i, size_t j);
+    int16_t get_value(size_t i, size_t j);
 
     size_t duplicates_col(size_t i);
     size_t duplicates_row(size_t i);
